requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
dj-database-url==2.3.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
cryptography==44.0.3
langchain==0.3.25
langchain-openai==0.3.18
langchain-chroma==0.2.6
chromadb==1.0.20
openai==1.78.1
jieba==0.42.1
pypdf==5.6.0
pandas==2.2.3
scipy==1.15.3
reportlab==4.4.0
channels[daphne]==4.2.0
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==11.1.0
